Fully remove Globals with the new blackboard SDK component
