+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/DriveTrainKt.kt
@@ -7,7 +7,6 @@ import com.qualcomm.robotcore.hardware.HardwareMap
 import org.firstinspires.ftc.robotcontroller.teamcode.GamepadButtons
 import org.firstinspires.ftc.robotcontroller.teamcode.HardwareMechanismKt
 import org.firstinspires.ftc.robotcore.external.Telemetry
-import org.firstinspires.ftc.teamcode.GlobalsKt
 import org.firstinspires.ftc.teamcode.misc.PoseKt
 import kotlin.math.abs
 import kotlin.math.max
@@ -19,7 +18,7 @@ class DriveTrainKt private constructor(hardwareMap: HardwareMap, data: InitData,
     private val driveMotors: Array<DcMotorEx> = createDriveMotors(hardwareMap)
     private val switch: DigitalChannel? = // optional hardware
         runCatching { hardwareMap.digitalChannel.get("switch") }.getOrNull()
-    private val referenceAngle = GlobalsKt.robotHeading // saved from auto, or 0 by default.
+    private val referenceAngle = data.referenceAngle // saved from auto, or our current heading by default.
     private val dashboardEnabled = data.dashboardEnabled
     private var currentSpeedMode = SPEEDS.NORMAL
     private var orientationMode = data.driveMode
@@ -103,7 +102,7 @@ class DriveTrainKt private constructor(hardwareMap: HardwareMap, data: InitData,
         NORMAL(0.65),
         FAST(0.80),
         SLOW(0.40),
-        CUSTOM_FTC_DASHBOARD(GlobalsKt.CUSTOM_FTC_DASHBOARD_SPEED)
+        CUSTOM_FTC_DASHBOARD(CUSTOM_FTC_DASHBOARD_SPEED)
     }
 
     private fun createDriveMotors(hardwareMap: HardwareMap): Array<DcMotorEx> {
@@ -128,5 +127,8 @@ class DriveTrainKt private constructor(hardwareMap: HardwareMap, data: InitData,
      */
     fun getSwitchState(): Boolean { return switch?.state ?: true }
 
-    companion object : HardwareMechanismSingletonManager<DriveTrainKt>(::DriveTrainKt)
+    companion object : HardwareMechanismSingletonManager<DriveTrainKt>(::DriveTrainKt) {
+        // This speed is designed to be set dynamically within FTC Dashboard.
+        @JvmField var CUSTOM_FTC_DASHBOARD_SPEED = 0.65
+    }
 }
